# pkg/api/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Cloud-Barista cm-centipede — Multi-cloud filesystem, object storage, and DBMS data migration subsystem.",
        "title": "cm-centipede API",
        "contact": {
            "name": "Cloud-Barista Community",
            "url": "https://github.com/cloud-barista/cm-centipede"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8085",
    "basePath": "/",
    "paths": {
        "/centipede/migration": {
            "get": {
                "description": "Returns a paginated list of migrations, optionally filtered by status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "migration"
                ],
                "summary": "List migrations (paginated)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status (pending|running|completed|failed|cancelled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20, max 100)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_ListResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_Migration"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a Migration record from the given plan and immediately starts execution asynchronously.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "migration"
                ],
                "summary": "Create and start a migration",
                "parameters": [
                    {
                        "description": "Migration creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.CreateMigrationReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_Migration"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    }
                }
            }
        },
        "/centipede/migration/all": {
            "get": {
                "description": "Returns every migration record ordered by created_at DESC, optionally filtered by status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "migration"
                ],
                "summary": "List all migrations without pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status (pending|running|completed|failed|cancelled)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-array_github_com_cloud-barista_cm-centipede_pkg_api_rest_model_Migration"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    }
                }
            }
        },
        "/centipede/migration/{migrationId}": {
            "get": {
                "description": "Returns the Migration record for the given ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "migration"
                ],
                "summary": "Get a single migration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Migration ID",
                        "name": "migrationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_Migration"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a migration record. Returns 409 if the migration is currently running — cancel it first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "migration"
                ],
                "summary": "Delete a migration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Migration ID",
                        "name": "migrationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    }
                }
            }
        },
        "/centipede/migration/{migrationId}/cancel": {
            "post": {
                "description": "Signals the running migration to stop after the current item finishes. Returns 404 if the migration is not currently running.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "migration"
                ],
                "summary": "Cancel a running migration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Migration ID",
                        "name": "migrationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_Migration"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    }
                }
            }
        },
        "/centipede/migration/{migrationId}/logs": {
            "get": {
                "description": "Returns paginated MigrationLog entries for the given migration ID, optionally filtered by status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "migration"
                ],
                "summary": "List logs for a migration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Migration ID",
                        "name": "migrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (success|failed|skipped)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20, max 100)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_ListResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_MigrationLog"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    }
                }
            }
        },
        "/centipede/migration/{migrationId}/retry": {
            "post": {
                "description": "Creates a new Migration that re-runs items from the original according to retryMode. DBMS destination tables are DROP-ped before re-execution.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "migration"
                ],
                "summary": "Retry a failed or cancelled migration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Original Migration ID",
                        "name": "migrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Retry options (retryMode: partial|full, default: partial)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.RetryReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_RetryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_cloud-barista_cm-centipede_dmdl_target-model.TargetDataMigrationModel": {
            "type": "object",
            "required": [
                "targetDataMigrationModel"
            ],
            "properties": {
                "targetDataMigrationModel": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.TargetDataMigrationProperty"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-array_github_com_cloud-barista_cm-centipede_pkg_api_rest_model_Migration": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.Migration"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_ListResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_Migration": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ListResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_Migration"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_ListResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_MigrationLog": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ListResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_MigrationLog"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_Migration": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.Migration"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_RetryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.RetryResponse"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.CreateMigrationReq": {
            "type": "object",
            "required": [
                "name",
                "plan"
            ],
            "properties": {
                "dbmsOnFailure": {
                    "description": "DBMSOnFailure decides what happens to a DBMS target when one of its\ndatabase items fails. It is named for the DBMS domain because only that\ndomain has an undo step: a filesystem or object storage transfer leaves\nwhatever it had already written, with or without this field.\n\n\"cleanup\" (the default) undoes the failed item. \"keep\" leaves it in place\nfor inspection — useful when the failure is in the data rather than in the\nconnection, since cleanup destroys the evidence.\n\nCancellation is not a failure and always cleans up, whatever this says.",
                    "type": "string",
                    "enum": [
                        "cleanup",
                        "keep"
                    ]
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "plan": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.TargetDataMigrationModel"
                }
            }
        },
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ListResponse-github_com_cloud-barista_cm-centipede_pkg_api_rest_model_Migration": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.Migration"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.Migration": {
            "type": "object",
            "required": [
                "name",
                "plan"
            ],
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "dbmsOnFailure": {
                    "description": "DBMSOnFailure is carried on the record rather than only on the request\nbecause execution is asynchronous and a retry re-reads the row: the policy\nhas to outlive the call that set it. Empty means DBMSOnFailureCleanup.",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "failedItems": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "plan": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.TargetDataMigrationModel"
                },
                "processedItems": {
                    "type": "integer"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "statusMessage": {
                    "type": "string"
                },
                "totalBytes": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "transferredBytes": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "validationDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ValidationDetailItem"
                    }
                },
                "validationMessage": {
                    "type": "string"
                },
                "validationStatus": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-centipede_pkg_api_rest_model.RetryReq": {
            "type": "object",
            "properties": {
                "retryMode": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}